requirements.txt
# Sito Consiglio di Amministrazione AI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.0
google-genai>=1.0
anthropic>=0.40.0
PyJWT>=2.8.0
